TSIS/end/requirements.txt
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
idna==2.10
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
virtualenv==16.6.0
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
gunicorn==20.1.0
